+++ CHANGELOG.md
@@ -17,6 +17,7 @@ None
 - Fix fringe_field_exit of `Dipole` is overwritten by `fringe_field` bug (see #99) (@cr-xu)
 - Fix error caused by mismatched devices on machines with CUDA GPUs (see #97) (@jank324)
 - Fix error raised when tracking a `ParameterBeam` through an active `BPM` (see #101) (@jank324)
+- Fix error in ASTRA beam import where the energy was set to `float64` instead of `float32` (see #111) (@jank324)
 
 ### 🐆 Other